alterando as migrations para rodar perfeitamente no meu local

+++ database/migrations/2020_02_05_123115_create_arquivos_table.php
@@ -22,8 +22,8 @@ class CreateArquivosTable extends Migration
             $table->boolean('versaoFinal')->nullable();
             $table->date('data')->nullable();
 
-            $table->integer('trabalhoId');
-            $table->integer('participanteId');
+            $table->unsignedBigInteger('trabalhoId');
+            $table->unsignedBigInteger('participanteId');
         });
     }
