+++ database/migrations/2020_02_05_123205_create_atividades_table.php
@@ -21,7 +21,7 @@ class CreateAtividadesTable extends Migration
             $table->boolean('versaoFinal');
             $table->date('data');
 
-            $table->integer('eventoId');
+            $table->unsignedBigInteger('eventoId');
         });
     }
