+++ database/migrations/2020_02_05_123453_create_coautors_table.php
@@ -18,8 +18,8 @@ class CreateCoautorsTable extends Migration
             $table->timestamps();
             $table->string('ordem');
 
-            $table->integer('autorId');
-            $table->integer('trabalhoId');
+            $table->unsignedBigInteger('autorId');
+            $table->unsignedBigInteger('trabalhoId');
         });
     }
