+++ database/migrations/2020_02_05_123503_create_parecers_table.php
@@ -18,8 +18,8 @@ class CreateParecersTable extends Migration
             $table->timestamps();
             $table->string('resultado');
 
-            $table->integer('revisorId');
-            $table->integer('trabalhoId');
+            $table->unsignedBigInteger('revisorId');
+            $table->unsignedBigInteger('trabalhoId');
         });
     }
