+++ database/migrations/2024_07_11_071852_create_area_tematica_pibac_avaliadors_table.php
@@ -16,10 +16,10 @@ class CreateAreaTematicaPibacAvaliadorsTable extends Migration
         Schema::create('area_tematica_pibac_avaliadors', function (Blueprint $table)
         {
             $table->bigIncrements('id');
-            $table->integer('area_tematica_pibac_id');
+            $table->unsignedBigInteger('area_tematica_pibac_id');
             $table->foreign('area_tematica_pibac_id')->references('id')->on('area_tematica_pibacs');
 
-            $table->integer('avaliador_id');
+            $table->unsignedBigInteger('avaliador_id');
             $table->foreign('avaliador_id')->references('id')->on('avaliadors');
 
             $table->timestamps();
